+++ src/pages/CRFdetail/models/crf_interview.js
@@ -1,0 +1,21 @@
+import { message } from "antd"
+
+const Model = {
+
+    namespace: "crf_interview",
+
+    state: {
+
+    },
+
+    reducers: {
+
+    },
+
+    effects: {
+
+    }
+
+}
+
+export default Model